+++ src/DependencyInjection/MisdPhoneNumberExtension.php
@@ -39,11 +39,15 @@ class MisdPhoneNumberExtension extends Extension
         }
         if ($config['serializer']['enabled']) {
             $loader->load('serializer.xml');
+
+            $container->setParameter('misd_phone_number.serializer.default_region', $config['serializer']['default_region']);
+            $container->setParameter('misd_phone_number.serializer.format', $config['serializer']['format']);
         }
         if ($config['validator']['enabled']) {
             $loader->load('validator.xml');
 
             $container->setParameter('misd_phone_number.validator.default_region', $config['validator']['default_region']);
+            $container->setParameter('misd_phone_number.validator.format', $config['validator']['format']);
         }
     }
 }